car_price_predictor/requirements.txt
Django>=3.2,<4.0
numpy
pandas
scikit-learn
shap
plotly
Pillow
pytesseract
fitz
langchain
langchain-text-splitters
langchain-community
langchain-ollama
joblib
```